requirements.txt
aider==0.2.4
aider-chat==0.18.0
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
asyncio==3.4.3
attrs==23.1.0
autogen==1.0.16
autopep8==2.0.4
backoff==2.2.1
blinker==1.6.3
cachetools==5.3.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
configparser==6.0.0
dataclasses-json==0.6.2
decorator==5.1.1
diskcache==5.6.3
distro==1.8.0
docopt==0.6.2
docutils==0.20.1
exceptiongroup==1.1.3
executing==2.0.1
FLAML==2.1.1
Flask==3.0.0
flask-requests==0.0.14
frozenlist==1.4.0
gevent==23.9.1
gitdb==4.0.11
GitPython==3.1.40
greenlet==3.0.1
grep-ast==0.2.4
grequests==0.7.0
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.1
hugchat==0.3.7
idna==3.4
importlib-metadata==6.8.0
ipython==8.17.2
itsdangerous==2.1.2
jaraco.classes==3.3.0
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
keyring==24.2.0
langchain==0.0.339
langsmith==0.0.66
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib-inline==0.1.6
mdurl==0.1.2
more-itertools==10.1.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
nh3==0.2.14
numpy==1.26.1
openai==1.3.4
packaging==23.2
pampy==0.3.0
pandas==2.1.2
parso==0.8.3
pathspec==0.11.2
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
plotly==5.18.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==13.0.0
pyautogen==0.1.14
pycodestyle==2.11.1
pycparser==2.21
pydantic==2.5.1
pydantic_core==2.14.3
pydeck==0.8.1b0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
readme-renderer==42.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==13.6.0
rpds-py==0.10.6
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
sklearn==0.0.post10
smmap==5.0.1
sniffio==1.3.0
sounddevice==0.4.6
soundfile==0.12.1
SQLAlchemy==2.0.23
stack-data==0.6.3
streamlit==1.28.0
streamlit-chat==0.1.1
streamlit-code-editor==0.1.10
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.13.0
tree-sitter==0.20.2
tree-sitter-languages==1.8.0
twine==4.0.2
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
urllib3==2.0.7
validators==0.22.0
wcwidth==0.2.9
Werkzeug==3.0.1
yarl==1.9.2
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
